test: synchronize realtime e2e readiness
