+++ src/web/src/test/e2e-ui/03-realtime-multiuser.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from "./_fixtures/community-fixture"
 import { tid } from "./_fixtures/testids"
-import { sendMessage, expectMessageVisible, composerEditable } from "./_fixtures/actions"
+import { sendMessage, expectMessageVisible, composerEditable, gotoAfterUserWsAuth } from "./_fixtures/actions"
 import { seedServer, seedChannel, seedJoinServer } from "./_fixtures/seed"
 
 // Journey 3 — multi-user realtime. Alice and Bob are both online in the same
@@ -53,8 +53,10 @@ test.describe.serial("multi-user realtime", () => {
   test("Bob sees Alice's typing indicator, which clears after her message", async ({ asUser }) => {
     const alice = await asUser("alice")
     const bob = await asUser("bob")
-    await alice.page.goto(`/c/channels/${serverId}/${channelId}`)
-    await bob.page.goto(`/c/channels/${serverId}/${channelId}`)
+    await Promise.all([
+      gotoAfterUserWsAuth(alice.page, `/c/channels/${serverId}/${channelId}`),
+      gotoAfterUserWsAuth(bob.page, `/c/channels/${serverId}/${channelId}`),
+    ])
     await bob.page.waitForURL(new RegExp(channelId), { timeout: 20_000 , waitUntil: "commit" })
 
     // Gate on Bob's WS subscription being LIVE before testing the typing
