+++ src/web/src/test/e2e-ui/14-forum-post-tags-presence.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect, userId } from "./_fixtures/community-fixture"
-import type { Page, WebSocket } from "@playwright/test"
 import { tid } from "./_fixtures/testids"
+import { gotoAfterUserWsAuth } from "./_fixtures/actions"
 import {
   seedServer,
   seedChannel,
@@ -10,38 +10,6 @@ import {
   seedDm,
   seedDmMessage,
 } from "./_fixtures/seed"
-
-async function gotoAfterUserWsAuth(page: Page, url: string): Promise<void> {
-  const frameHandlers = new Map<WebSocket, (event: { payload: string | Buffer }) => void>()
-  let cleanup = () => {}
-  const authenticated = new Promise<void>((resolve, reject) => {
-    const timer = setTimeout(() => reject(new Error("user WebSocket did not authenticate")), 20_000)
-    const onWebSocket = (socket: WebSocket) => {
-      const onFrame = (event: { payload: string | Buffer }) => {
-        try {
-          const message = JSON.parse(event.payload.toString()) as { type?: string }
-          if (message.type !== "auth.ok") return
-          cleanup()
-          resolve()
-        } catch {}
-      }
-      frameHandlers.set(socket, onFrame)
-      socket.on("framereceived", onFrame)
-    }
-    cleanup = () => {
-      clearTimeout(timer)
-      page.off("websocket", onWebSocket)
-      for (const [socket, handler] of frameHandlers) socket.off("framereceived", handler)
-    }
-    page.on("websocket", onWebSocket)
-  })
-
-  try {
-    await Promise.all([page.goto(url, { waitUntil: "commit" }), authenticated])
-  } finally {
-    cleanup()
-  }
-}
 
 // Journey 14 — forum-post notify scope, per-post tags, post-card participant
 // avatars, and DM presence stability. Covers the batch that made forum posts
@@ -154,6 +122,13 @@ test.describe.serial("DM presence stability on refresh", () => {
     await alice.page.waitForURL(/\/c\/me/, { timeout: 20_000, waitUntil: "commit" })
 
     const bob = await asUser("bob")
+    await expect.poll(async () => {
+      const response = await bob.page.request.get(`/api/community/servers/${serverId}/presence`)
+      if (!response.ok()) return false
+      const body = await response.json() as { online?: string[] }
+      return body.online?.includes(userId("alice")) ?? false
+    }, { timeout: 20_000 }).toBe(true)
+
     await bob.page.goto(`/c/me/${dmId}`)
     await bob.page.waitForURL(new RegExp(dmId), { timeout: 20_000, waitUntil: "commit" })
