+++ src/web/src/test/e2e-ui/_fixtures/actions.ts
@@ -1,5 +1,37 @@
-import { type Page, expect } from "@playwright/test"
+import { type Page, type WebSocket, expect } from "@playwright/test"
 import { tid } from "./testids"
+
+export async function gotoAfterUserWsAuth(page: Page, url: string): Promise<void> {
+  const frameHandlers = new Map<WebSocket, (event: { payload: string | Buffer }) => void>()
+  let cleanup = () => {}
+  const authenticated = new Promise<void>((resolve, reject) => {
+    const timer = setTimeout(() => reject(new Error("user WebSocket did not authenticate")), 20_000)
+    const onWebSocket = (socket: WebSocket) => {
+      const onFrame = (event: { payload: string | Buffer }) => {
+        try {
+          const message = JSON.parse(event.payload.toString()) as { type?: string }
+          if (message.type !== "auth.ok") return
+          cleanup()
+          resolve()
+        } catch {}
+      }
+      frameHandlers.set(socket, onFrame)
+      socket.on("framereceived", onFrame)
+    }
+    cleanup = () => {
+      clearTimeout(timer)
+      page.off("websocket", onWebSocket)
+      for (const [socket, handler] of frameHandlers) socket.off("framereceived", handler)
+    }
+    page.on("websocket", onWebSocket)
+  })
+
+  try {
+    await Promise.all([page.goto(url, { waitUntil: "commit" }), authenticated])
+  } finally {
+    cleanup()
+  }
+}
 
 // Reusable UI action helpers built on the canonical testids. Journeys compose
 // these so the real user path (click → type → assert) stays readable and the
